+++ src/views/BasicData/Airport/index.vue
@@ -62,7 +62,7 @@ const crudSchemas = reactive<CrudSchema[]>([
   //   },
   {
     field: 'index',
-    label: t('tableDemo.index'),
+    label: t('common.index'),
     type: 'index',
     search: {
       hidden: true
@@ -121,10 +121,10 @@ const crudSchemas = reactive<CrudSchema[]>([
               }
             >
               {data.importance === 1
-                ? t('tableDemo.important')
+                ? t('common.important')
                 : data.importance === 2
-                ? t('tableDemo.good')
-                : t('tableDemo.commonly')}
+                ? t('common.good')
+                : t('common.commonly')}
             </ElTag>
           )
         }
@@ -159,7 +159,7 @@ const crudSchemas = reactive<CrudSchema[]>([
   {
     field: 'action',
     width: '260px',
-    label: t('tableDemo.action'),
+    label: t('common.action'),
     search: {
       hidden: true
     },
@@ -175,13 +175,13 @@ const crudSchemas = reactive<CrudSchema[]>([
           return (
             <>
               <ElButton type="primary" onClick={() => action(data.row, 'edit')}>
-                {t('exampleDemo.edit')}
+                {t('common.edit')}
               </ElButton>
               <ElButton type="success" onClick={() => action(data.row, 'detail')}>
-                {t('exampleDemo.detail')}
+                {t('common.detail')}
               </ElButton>
               <ElButton type="danger" onClick={() => delData(data.row)}>
-                {t('exampleDemo.del')}
+                {t('common.del')}
               </ElButton>
             </>
           )
@@ -201,7 +201,7 @@ const currentRow = ref<TableData | null>(null)
 const actionType = ref('')
 
 const AddAction = () => {
-  dialogTitle.value = t('exampleDemo.add')
+  dialogTitle.value = t('common.add')
   currentRow.value = null
   dialogVisible.value = true
   actionType.value = ''
@@ -219,7 +219,7 @@ const delData = async (row: TableData | null) => {
 }
 
 const action = (row: TableData, type: string) => {
-  dialogTitle.value = t(type === 'edit' ? 'exampleDemo.edit' : 'exampleDemo.detail')
+  dialogTitle.value = t(type === 'edit' ? 'common.edit' : 'common.detail')
   actionType.value = type
   currentRow.value = row
   dialogVisible.value = true
@@ -253,9 +253,9 @@ const save = async () => {
     <Search :schema="allSchemas.searchSchema" @search="setSearchParams" @reset="setSearchParams" />
 
     <div class="mb-10px">
-      <ElButton type="primary" @click="AddAction">{{ t('exampleDemo.add') }}</ElButton>
+      <ElButton type="primary" @click="AddAction">{{ t('common.add') }}</ElButton>
       <ElButton :loading="delLoading" type="danger" @click="delData(null)">
-        {{ t('exampleDemo.del') }}
+        {{ t('common.del') }}
       </ElButton>
     </div>
 
@@ -288,9 +288,9 @@ const save = async () => {
 
     <template #footer>
       <ElButton v-if="actionType !== 'detail'" type="primary" :loading="saveLoading" @click="save">
-        {{ t('exampleDemo.save') }}
+        {{ t('common.save') }}
       </ElButton>
-      <ElButton @click="dialogVisible = false">{{ t('dialogDemo.close') }}</ElButton>
+      <ElButton @click="dialogVisible = false">{{ t('common.close') }}</ElButton>
     </template>
   </Dialog>
 </template>