+++ src/views/Interference/components/AddDialog.vue
@@ -63,7 +63,7 @@ const save = async () => {
         {{ t('common.back') }}
       </ElButton>
       <ElButton type="primary" :loading="loading" @click="save">
-        {{ t('exampleDemo.save') }}
+        {{ t('common.save') }}
       </ElButton>
     </template>
   </Dialog>