+++ src/views/Interference/index.vue
@@ -65,7 +65,7 @@ const { t } = useI18n()
 const crudSchemas = reactive<CrudSchema[]>([
   {
     field: 'index',
-    label: t('tableDemo.index'),
+    label: t('common.index'),
     type: 'index',
     search: {
       hidden: true
@@ -166,10 +166,10 @@ const crudSchemas = reactive<CrudSchema[]>([
               }
             >
               {data.importance === 1
-                ? t('tableDemo.important')
+                ? t('common.important')
                 : data.importance === 2
-                ? t('tableDemo.good')
-                : t('tableDemo.commonly')}
+                ? t('common.good')
+                : t('common.commonly')}
             </ElTag>
           )
         }
@@ -183,7 +183,7 @@ const crudSchemas = reactive<CrudSchema[]>([
   {
     field: 'action',
     width: '260px',
-    label: t('tableDemo.action'),
+    label: t('common.action'),
     search: {
       hidden: true
     },
@@ -199,13 +199,13 @@ const crudSchemas = reactive<CrudSchema[]>([
           return (
             <>
               <ElButton type="primary" onClick={() => action(data.row, 'edit')}>
-                {t('exampleDemo.edit')}
+                {t('common.edit')}
               </ElButton>
               <ElButton type="success" onClick={() => action(data.row, 'detail')}>
-                {t('exampleDemo.detail')}
+                {t('common.detail')}
               </ElButton>
               <ElButton type="danger" onClick={() => delData(data.row)}>
-                {t('exampleDemo.del')}
+                {t('common.del')}
               </ElButton>
             </>
           )
@@ -243,9 +243,9 @@ const action = (row: TableData, type: string) => {
     <Search :schema="allSchemas.searchSchema" @search="setSearchParams" @reset="setSearchParams" />
 
     <div class="mb-10px">
-      <ElButton type="primary" @click="AddAction">{{ t('exampleDemo.add') }}</ElButton>
+      <ElButton type="primary" @click="AddAction">{{ t('common.add') }}</ElButton>
       <ElButton :loading="delLoading" type="danger" @click="delData(null)">
-        {{ t('exampleDemo.del') }}
+        {{ t('common.del') }}
       </ElButton>
     </div>